+++ packages/tswebextension/CHANGELOG.md
@@ -16,6 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `$cookie` modifier in MV3 via browser.cookies API and content-script.
 
 
+## [0.3.12] - 2023-08-23
+
+### Fixed
+- Incorrect order of `onBeforeRequest` handlers.
+- Blocked csp reports do not increment the blocked requests counter.
+
+
 ## [0.3.11] - 2023-08-21
 
 ### Changed